keyvisual: update decorator and matrix package
